--- guess-game-server/src/main/java/guess/domain/source/contentful/ContentfulEventResponse.java
@@ -1,4 +1,0 @@
-package guess.domain.source.contentful;
-
-public class ContentfulEventResponse extends ContentfulResponse<ContentfulEvent> {
-}
