--- guess-game-server/src/main/java/guess/domain/source/contentful/ContentfulLocaleResponse.java
@@ -1,4 +1,0 @@
-package guess.domain.source.contentful;
-
-public class ContentfulLocaleResponse extends ContentfulResponse<ContentfulLocale> {
-}
