--- guess-game-server/src/main/java/guess/domain/source/contentful/ContentfulTalkResponse.java
@@ -1,4 +1,0 @@
-package guess.domain.source.contentful;
-
-public class ContentfulTalkResponse extends ContentfulResponse<ContentfulTalk> {
-}
