+++ guess-game-server/src/main/java/guess/domain/source/contentful/event/ContentfulEvent.java
@@ -1,4 +1,4 @@
-package guess.domain.source.contentful;
+package guess.domain.source.contentful.event;
 
 public class ContentfulEvent {
     private ContentfulEventFields fields;
