+++ guess-game-server/src/main/java/guess/domain/source/contentful/event/ContentfulEventResponse.java
@@ -1,0 +1,6 @@
+package guess.domain.source.contentful.event;
+
+import guess.domain.source.contentful.ContentfulResponse;
+
+public class ContentfulEventResponse extends ContentfulResponse<ContentfulEvent> {
+}
