+++ guess-game-server/src/main/java/guess/domain/source/contentful/eventtype/ContentfulEventTypeFields.java
@@ -1,4 +1,4 @@
-package guess.domain.source.contentful;
+package guess.domain.source.contentful.eventtype;
 
 public class ContentfulEventTypeFields {
     private String eventName;
