+++ guess-game-server/src/main/java/guess/domain/source/contentful/eventtype/ContentfulEventTypeResponse.java
@@ -1,0 +1,6 @@
+package guess.domain.source.contentful.eventtype;
+
+import guess.domain.source.contentful.ContentfulResponse;
+
+public class ContentfulEventTypeResponse extends ContentfulResponse<ContentfulEventType> {
+}
