+++ guess-game-server/src/main/java/guess/domain/source/contentful/locale/ContentfulLocale.java
@@ -1,4 +1,4 @@
-package guess.domain.source.contentful;
+package guess.domain.source.contentful.locale;
 
 public class ContentfulLocale {
     private String code;
