+++ guess-game-server/src/main/java/guess/domain/source/contentful/locale/ContentfulLocaleResponse.java
@@ -1,0 +1,6 @@
+package guess.domain.source.contentful.locale;
+
+import guess.domain.source.contentful.ContentfulResponse;
+
+public class ContentfulLocaleResponse extends ContentfulResponse<ContentfulLocale> {
+}
