+++ guess-game-server/src/main/java/guess/domain/source/contentful/speaker/ContentfulSpeakerResponse.java
@@ -1,0 +1,6 @@
+package guess.domain.source.contentful.speaker;
+
+import guess.domain.source.contentful.ContentfulResponse;
+
+public class ContentfulSpeakerResponse extends ContentfulResponse<ContentfulSpeaker> {
+}
