+++ guess-game-server/src/main/java/guess/domain/source/contentful/talk/ContentfulTalkResponse.java
@@ -1,0 +1,6 @@
+package guess.domain.source.contentful.talk;
+
+import guess.domain.source.contentful.ContentfulResponse;
+
+public class ContentfulTalkResponse extends ContentfulResponse<ContentfulTalk> {
+}
